+++ R/RandomForest.R
@@ -118,7 +118,7 @@ fitRandomForest <- function(population, plpData, param, search='grid', quiet=F,
   ParallelLogger::logTrace('Mapping R data to python')
   #x <- toSparsePython2(plpData,population, map=NULL)
   prediction <- population
-  x <- toSparseM(plpData,population,map=NULL, temporal = F)
+  x <- toSparseM(plpData,population,map=NULL)
 
   ParallelLogger::logInfo('Sourcing python code')
   reticulate::source_python(system.file(package='PatientLevelPrediction','python','randomForestFunctions.py'), envir = e)